+++ R/release.r
@@ -16,6 +16,10 @@
 #'   \item Draft an email to the CRAN maintainer.
 #' }
 #'
+#' You can also add arbitrary extra questions by defining an (un-exported)
+#' function called \code{release_questions()} that returns a character vector
+#' of additional questions to ask.
+#'
 #' You also need to read the CRAN repository policy at
 #' \url{http://cran.r-project.org/web/packages/policies.html} and make
 #' sure you're in line with the policies. \code{release} tries to automate as
@@ -91,6 +95,14 @@ release <- function(pkg = ".", check = TRUE) {
 
     if (yesno(msg))
       return(invisible())
+  }
+
+  release_questions <- pkg_env(pkg)$release_questions
+  if (!is.null(release_questions)) {
+    questions <- release_questions()
+    for (question in questions) {
+      if (yesno(question)) return(invisible())
+    }
   }
 
   if (yesno("Ready to upload?"))